+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### ✅ Added
 - Picture-in-Picture support
 - Livestream Player
+- Call stats report
+
+### 🔄 Changed
+- Factory method for creating `LocalParticipantViewModifier`
 
 # [0.3.0](https://github.com/GetStream/stream-video-swift/releases/tag/0.3.0)
 _August 25, 2023_